requirements.txt
pandas
quickfix
kafka-python
scikit-learn
